dcp-845 service test, fixes
