+++ requirements-dev.in
@@ -1,4 +1,5 @@
 -c requirements.txt
+assertpy
 coverage
 pylint
 pandas
